egl: Use SConscript for Windows build.
Fix several portability issues and add SConscript for Windows build.

+++ src/egl/main/eglstring.c
@@ -11,7 +11,7 @@ char *
 _eglstrdup(const char *s)
 {
    if (s) {
-      int l = strlen(s);
+      size_t l = strlen(s);
       char *s2 = malloc(l + 1);
       if (s2)
          strcpy(s2, s);
